requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
appnope==0.1.4
    # via ipykernel
apptools==5.2.1
    # via
    #   envisage
    #   mayavi
asttokens==2.4.1
    # via stack-data
build==1.2.1
    # via pip-tools
certifi==2024.2.2
    # via
    #   requests
    #   sentry-sdk
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   pip-tools
    #   wandb
comm==0.2.2
    # via ipykernel
configobj==5.0.8
    # via apptools
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
cython==3.0.11
    # via ptitprince
debugpy==1.8.1
    # via ipykernel
decorator==5.1.1
    # via ipython
distlib==0.3.8
    # via virtualenv
docker-pycreds==0.4.0
    # via wandb
envisage==7.0.3
    # via mayavi
exceptiongroup==1.2.1
    # via ipython
executing==2.0.1
    # via stack-data
filelock==3.14.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   virtualenv
fonttools==4.52.4
    # via matplotlib
fsspec==2024.5.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via wandb
huggingface-hub==0.23.2
    # via
    #   tokenizers
    #   transformers
identify==2.5.36
    # via pre-commit
idna==3.7
    # via requests
ipykernel==6.29.4
    # via LCA-normative-modelling (pyproject.toml)
ipython==8.24.0
    # via ipykernel
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via torch
joblib==1.4.2
    # via
    #   nilearn
    #   scikit-learn
jupyter-client==8.6.2
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.5
    # via matplotlib
lxml==5.2.2
    # via nilearn
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.0
    # via
    #   LCA-normative-modelling (pyproject.toml)
    #   nichord
    #   ptitprince
    #   pysurfer
    #   seaborn
    #   vtk
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mayavi==4.8.2
    # via pysurfer
mpmath==1.3.0
    # via sympy
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.3
    # via torch
neurocombat==0.2.12
    # via LCA-normative-modelling (pyproject.toml)
nibabel==5.2.1
    # via
    #   nilearn
    #   pysurfer
nichord==0.3.0
    # via LCA-normative-modelling (pyproject.toml)
nilearn==0.10.4
    # via
    #   LCA-normative-modelling (pyproject.toml)
    #   nichord
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   mayavi
    #   nibabel
    #   nichord
    #   nilearn
    #   pandas
    #   patsy
    #   ptitprince
    #   pysurfer
    #   scikit-learn
    #   scipy
    #   seaborn
    #   statsmodels
    #   torchvision
    #   transformers
packaging==24.0
    # via
    #   build
    #   huggingface-hub
    #   ipykernel
    #   matplotlib
    #   mayavi
    #   nibabel
    #   nilearn
    #   plotly
    #   statsmodels
    #   transformers
pandas==2.2.2
    # via
    #   LCA-normative-modelling (pyproject.toml)
    #   nichord
    #   nilearn
    #   seaborn
    #   statsmodels
parso==0.8.4
    # via jedi
patsy==0.5.6
    # via statsmodels
pexpect==4.9.0
    # via ipython
pillow==10.3.0
    # via
    #   matplotlib
    #   nichord
    #   torchvision
pip-tools==7.4.1
    # via LCA-normative-modelling (pyproject.toml)
platformdirs==4.2.2
    # via
    #   jupyter-core
    #   virtualenv
    #   wandb
plotly==5.22.0
    # via LCA-normative-modelling (pyproject.toml)
pre-commit==3.7.1
    # via LCA-normative-modelling (pyproject.toml)
prompt-toolkit==3.0.44
    # via ipython
protobuf==4.25.3
    # via wandb
psutil==5.9.8
    # via
    #   ipykernel
    #   wandb
ptitprince==0.2.5
    # via LCA-normative-modelling (pyproject.toml)
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyface==8.0.0
    # via
    #   envisage
    #   mayavi
    #   traitsui
pygments==2.18.0
    # via
    #   ipython
    #   mayavi
pyhamcrest==2.1.0
    # via ptitprince
pyparsing==3.1.2
    # via matplotlib
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pysurfer==0.11.2
    # via LCA-normative-modelling (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   pre-commit
    #   transformers
    #   wandb
pyzmq==26.0.3
    # via
    #   ipykernel
    #   jupyter-client
regex==2024.5.15
    # via transformers
requests==2.32.2
    # via
    #   huggingface-hub
    #   nilearn
    #   transformers
    #   wandb
ruff==0.4.5
    # via LCA-normative-modelling (pyproject.toml)
safetensors==0.4.3
    # via transformers
scikit-learn==1.5.0
    # via
    #   LCA-normative-modelling (pyproject.toml)
    #   nilearn
scipy==1.13.1
    # via
    #   nichord
    #   nilearn
    #   ptitprince
    #   pysurfer
    #   scikit-learn
    #   statsmodels
seaborn==0.13.2
    # via
    #   LCA-normative-modelling (pyproject.toml)
    #   ptitprince
sentry-sdk==2.3.1
    # via wandb
setproctitle==1.3.3
    # via wandb
six==1.16.0
    # via
    #   asttokens
    #   configobj
    #   docker-pycreds
    #   patsy
    #   python-dateutil
smmap==5.0.1
    # via gitdb
stack-data==0.6.3
    # via ipython
statsmodels==0.14.2
    # via LCA-normative-modelling (pyproject.toml)
sympy==1.12
    # via torch
tenacity==8.3.0
    # via plotly
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.19.1
    # via transformers
tomli==2.0.1
    # via
    #   build
    #   pip-tools
torch==2.3.0
    # via
    #   LCA-normative-modelling (pyproject.toml)
    #   torchaudio
    #   torchvision
torchaudio==2.3.0
    # via LCA-normative-modelling (pyproject.toml)
torchvision==0.18.0
    # via LCA-normative-modelling (pyproject.toml)
tornado==6.4
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.66.4
    # via
    #   huggingface-hub
    #   transformers
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
traits==6.4.3
    # via
    #   apptools
    #   envisage
    #   mayavi
    #   pyface
    #   traitsui
traitsui==8.0.0
    # via
    #   apptools
    #   envisage
    #   mayavi
transformers==4.41.1
    # via LCA-normative-modelling (pyproject.toml)
typing-extensions==4.12.0
    # via
    #   huggingface-hub
    #   ipython
    #   torch
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via
    #   requests
    #   sentry-sdk
virtualenv==20.26.2
    # via pre-commit
vtk==9.3.0
    # via mayavi
wandb==0.17.0
    # via LCA-normative-modelling (pyproject.toml)
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
